// mesh-control-plane/mesh-ui/src/api/openapi/types.gen.ts
export type V1CreateUserRequest = {
  name?: string
  displayName?: string
  email?: string
  pictureUrl?: string
}

export type V1CreateUserResponse = {
  user?: V1User
}

export type V1DebugCreateNodeRequest = {
  user?: string
  key?: string
  name?: string
  routes?: Array<string>
}

export type V1DebugCreateNodeResponse = {
  node?: V1Node
}

export type V1DeleteApiKeyResponse = {
  [key: string]: unknown
}

export type V1DeleteNodeResponse = {
  [key: string]: unknown
}

export type V1DeleteUserResponse = {
  [key: string]: unknown
}

export type V1ExpireApiKeyRequest = {
  prefix?: string
}

export type V1ExpireApiKeyResponse = {
  [key: string]: unknown
}

export type V1ExpireNodeResponse = {
  node?: V1Node
}

export type V1ExpirePreAuthKeyRequest = {
  user?: string
  key?: string
}

export type V1ExpirePreAuthKeyResponse = {
  [key: string]: unknown
}

export type V1GetNodeResponse = {
  node?: V1Node
}

export type V1GetPolicyResponse = {
  policy?: string
  updatedAt?: string
}

export type V1ListApiKeysResponse = {
  apiKeys?: Array<V1ApiKey>
}

export type V1ListNodesResponse = {
  nodes?: Array<V1Node>
}

export type V1ListPreAuthKeysResponse = {
  preAuthKeys?: Array<V1PreAuthKey>
}

export type V1ListUsersResponse = {
  users?: Array<V1User>
}
// ...
export type V1SetPolicyResponse = {
  policy?: string
  updatedAt?: string
}

export type V1SetTagsResponse = {
  node?: V1Node
}

export type V1User = {
  id?: string
  name?: string
  createdAt?: string
  displayName?: string
  email?: string
  providerId?: string
  provider?: string
  profilePicUrl?: string
}